+++ R-autofun/plot-and-jitter.R
@@ -28,10 +28,10 @@ if (plotit && pedometrics::isNumint(k / 10)) {
     conf0 = conf0[, 2:3], y_max0 = y_max0, y.max = y.max, x_max0 = x_max0, x.max = x.max, 
     best.energy = best_energy, best.k = best_k, MOOP = MOOP)
 }
-new_conf <- spJitter(
-  points = old_conf, candi = candi, x.max = x.max, x.min = x.min, y.max = y.max, y.min = y.min,
-  #finite = finite,
-  cellsize = cellsize, which.point = wp)
+new_conf <- 
+  spJitter(
+    points = old_conf, candi = candi, x.max = x.max, x.min = x.min, y.max = y.max, y.min = y.min,
+    cellsize = cellsize, which.point = wp, verbose = verbose)
 # if (acceptance$by == "iterations") {
 #   x.max <- x_max0 - (k / iterations) * (x_max0 - x.min)
 #   y.max <- y_max0 - (k / iterations) * (y_max0 - y.min)